added methods, class, crud and exception.

+++ src/main/java/br/com/serratec/project/controller/ClientController.java
@@ -3,6 +3,7 @@ package br.com.serratec.project.controller;
 import br.com.serratec.project.dto.ClientDto;
 import br.com.serratec.project.service.ClientService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
@@ -18,7 +19,7 @@ public class ClientController {
     }
 
     @GetMapping("/{id}")
-    public ClientDto getClientById(@PathVariable Integer id) {
+    public ClientDto getClientById(@PathVariable Long id) {
         return clientService.searchById(id);
     }
 
@@ -28,12 +29,12 @@ public class ClientController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteCar(@PathVariable Integer id)  {
+    public void deleteCar(@PathVariable Long id)  {
         clientService.delete(id);
     }
 
     @PutMapping("{id}")
-    public void updateCar(@PathVariable int id, @RequestBody ClientDto car)  {
+    public void updateCar(@PathVariable Long id, @RequestBody ClientDto car)  {
         clientService.update(id, car);
     }
