+++ src/main/java/br/com/serratec/project/repository/BudgetRepository.java
@@ -5,5 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BudgetRepository extends JpaRepository<Budget, Integer> {
+public interface BudgetRepository extends JpaRepository<Budget, Long> {
+
 }
